+++ src/main/kotlin/no/nav/sokos/skattekort/infrastructure/skatteetaten/bestillskattekort/BestillSkattekortRequest.kt
@@ -3,6 +3,7 @@ package no.nav.sokos.skattekort.infrastructure.skatteetaten.bestillskattekort
 import kotlinx.serialization.Serializable
 
 import no.nav.sokos.skattekort.module.person.Personidentifikator
+import no.nav.sokos.skattekort.person.Personidentifikator
 
 // Data classes
 @Serializable
@@ -66,7 +67,7 @@ fun bestillSkattekortRequest(
             ),
     )
 
-fun bestillOppdateringRequest(inntektsaar: Short): BestillSkattekortRequest =
+fun bestillOppdateringRequest(inntektsaar: Int): BestillSkattekortRequest =
     BestillSkattekortRequest(
         inntektsaar = inntektsaar.toString(),
         bestillingstype = "HENT_KUN_ENDRING",
